% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctList.R
\name{ctList}
\alias{ctList}
\title{Country and Territory List}
\usage{
ctList()
}
\description{
Generate a list of all the countries and territories in the list.
}
\author{
Joel Le Forestier (joel.leforestier@mail.utoronto.ca)
}
